Don't merge PR unless it has "awaiting merge" label
Change the behavior for when deciding when miss-islington
should merge the PR.

Previously it checks if the PR has one core dev who approved the PR.
But that is unreliable because the review may be dismissed later,
or another core dev might decide not to backport the PR anymore.

The "awaiting merge" label is a better indicator that the PR
is ready for merge.

Adjust tests.

Fixes https://github.com/python/miss-islington/issues/62